+++ app/src/main/java/com/adammcneilly/bloom/theme/Color.kt
@@ -1,0 +1,9 @@
+@file:Suppress("MagicNumber")
+
+package com.adammcneilly.bloom.theme
+
+import androidx.compose.ui.graphics.Color
+
+val Purple200 = Color(0xFFBB86FC)
+val Purple500 = Color(0xFF6200EE)
+val Teal200 = Color(0xFF03DAC5)
