--- app/src/main/java/template/theme/Color.kt
@@ -1,9 +1,0 @@
-@file:Suppress("MagicNumber")
-
-package template.theme
-
-import androidx.compose.ui.graphics.Color
-
-val Purple200 = Color(0xFFBB86FC)
-val Purple500 = Color(0xFF6200EE)
-val Teal200 = Color(0xFF03DAC5)
